agenticai-4/pyproject.toml
[project]
name = "agenticai-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "langgraph-api>=0.2.102",
    "langgraph-cli[inmem]>=0.3.6",
    "langgraph-runtime-inmem>=0.6.8",
    "langsmith>=0.4.8",
    "python-dotenv>=1.1.1",
]
